CQv2: Optimise the per-queue store
This removes the use of delayed_write and instead uses an
internal buffer that's also used as a cache, similar to how
the index works. The offset and size of messages in the file
are calculated using erlang:external_size/1 and as a result
the files may be a little bigger than before, but they should
not be significantly, especially considering messages are
mostly made of atoms and binaries.

The performance is boosted by around 10% on my machine
as a result of these changes.